self-review: drop two invariants the reject path cannot reach, and say why cardinal keeps its own check
Three findings from the review panel on the previous commit.

The zero-interval sweep case still carried `knots-non-decreasing` and
`interval-count` alongside the new `must_reject`. Neither can run: `run_case`
returns on the exception, and on a return it reports `must-reject:returned`
before the check list is built. They are dropped, which makes
`_interval_count`'s own docstring wrong where it says it is "the one that
separates them at n = 0" -- it no longer sees `n = 0` at all -- so that sentence
is corrected too. This is the project's dominant defect class sitting in code
the previous commit edited.

`create_cardinal_knots` keeps its own copy of the same check rather than
delegating to `_validate_knot_input`, and nothing said why. It derives its
domain from the count, so a zero reaches the validator as `domain=(0, 0)` and
comes back as "domain[0] must be less than domain[1]", naming a parameter the
caller never passed. Verified by calling the validator with those arguments. A
note now records it, and the changelog no longer claims the validator applies
the rule to all three -- it applies it to the two that call it.

The validator's `Args:` entry spelled the type `int_`.

+++ src/pantr/bspline/_bspline_knots.py
@@ -604,7 +604,7 @@ def _validate_knot_input(
     """Validate input parameters for knot vector generation.
 
     Args:
-        num_intervals (int_): Number of intervals in the domain. Must be at least 1:
+        num_intervals (int): Number of intervals in the domain. Must be at least 1:
             a mesh of zero intervals has no cell, so nothing is defined over it.
         degree (int): B-spline degree.
         continuity (int): Continuity level at interior knots.
